8/3 작업사항
- 플레이어 회피 동작 수정 ( 바라보는 방향을 기준으로 하도록 설정 )
- 캐릭터 무한 회전 버그 수정 ( 리지드바디의 프리즈 로테이션 y축 잠금 )
- 튜토리얼 UI 추가 및 HP UI 추가

+++ Assets/Scripts/Interfaces.cs
@@ -12,5 +12,5 @@ public class Interfaces : MonoBehaviour
 
 interface IDamage
 {
-    void Damage(int point);
+    void Damage(float point);
 }
